requirements.txt
python-telegram-bot
cloudmersive_ocr_api_client